* commit 'a954d74b4305223abc56ecaa6d098fd947bfea5e':
  cwd return bug
  fixed cast warnings
  removed temp files
  NLST fix

+++ dir.c
@@ -7,11 +7,11 @@
 #include "dir.h"
 #include <sys/stat.h>
 
-/* 
-   Arguments: 
+/*
+   Arguments:
       fd - a valid open file descriptor. This is not checked for validity
            or for errors with it is used.
-      directory - a pointer to a null terminated string that names a 
+      directory - a pointer to a null terminated string that names a
                   directory
 
    Returns
@@ -19,28 +19,28 @@
          to read it.
       -2 insufficient resources to perform request
 
- 
+
    This function takes the name of a directory and lists all the regular
-   files and directoies in the directory. 
- 
+   files and directoies in the directory.
+
 
  */
 
 int listFiles(int fd, char * directory) {
 
   // Get resources to see if the directory can be opened for reading
-  
+
   DIR * dir = NULL;
-  
+
   dir = opendir(directory);
   if (!dir) return -1;
-  
+
   // Setup to read the directory. When printing the directory
-  // only print regular files and directories. 
+  // only print regular files and directories.
 
   struct dirent *dirEntry;
   int entriesPrinted = 0;
-  
+
   for (dirEntry = readdir(dir);
        dirEntry;
        dirEntry = readdir(dir)) {
@@ -50,7 +50,7 @@ int listFiles(int fd, char * directory) {
       // This call really should check the return value
       stat(dirEntry->d_name, &buf);
 
-	dprintf(fd, "F    %-20s     %d\n", dirEntry->d_name, buf.st_size);
+	dprintf(fd, "F    %-20s     %d\n", dirEntry->d_name, (int) buf.st_size);
     } else if (dirEntry->d_type == DT_DIR) { // Directory
       dprintf(fd, "D        %s\n", dirEntry->d_name);
     } else {
@@ -58,10 +58,8 @@ int listFiles(int fd, char * directory) {
     }
     entriesPrinted++;
   }
-  
+
   // Release resources
   closedir(dir);
   return entriesPrinted;
 }
-
-
